fixed build issues with moving files around
