--- ZenJourney/backend/src/main/kotlin/com/backend/application/response/ListResponse.kt
@@ -1,6 +1,0 @@
-package com.backend.application.response
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ListResponse<T>(val list: List<T>)
